+++ src/main/server/utility/index.ts
@@ -1,3 +1,4 @@
 export * from './hash.js';
 export * from './password.js';
 export * from './reconnect.js';
+import './version.js';